requirements.txt
flask==3.0.2
flask-restful==0.3.10
tensorflow==2.16.1
pillow==10.2.0
werkzeug==3.0.1
flask-cors==4.0.0
numpy==1.24.3